Remove obsolete MapTaskRunner.
